Feat: UserGo uint64 variables checked, Also methods Find and Update parameters updated in uint64, also updated user displayed modified.
Receipt.go uses contents and amount and client id, initialized variables checked but stills has a problem in the gorm database implementation. Also uint parameters in the uint64 deleteAReceipt method uses uint64
Client.Go uint64 variables updated. Users controller doesn't parses any more for FindUserByID user related to the modification with the types. TokenID Validation also doesn't uses parsing any more.

Error: modification in getClients end point wich didn't called the correct one.
Create receipt doesnt receives a uid any more, due the receipt implementation didn't worked as expected.

+++ api/controllers/users_controller.go
@@ -69,7 +69,7 @@ func (server *Server) GetUser(w http.ResponseWriter, r *http.Request) {
 	}
 
 	user := models.User{}
-	userGotten, err := user.FindUserByID(server.DB, uint32(uid))
+	userGotten, err := user.FindUserByID(server.DB, uid)
 
 	if err != nil {
 		responses.ERROR(w, http.StatusBadRequest, err)
@@ -108,7 +108,7 @@ func (server *Server) UpdateUser(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	if tokenID != uint32(uid) {
+	if tokenID != uid {
 		responses.ERROR(w, http.StatusUnauthorized, errors.New(http.StatusText(http.StatusUnauthorized)))
 		return
 	}
@@ -121,7 +121,7 @@ func (server *Server) UpdateUser(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	updateUser, err := user.UpdateAUser(server.DB, uint32(uid))
+	updateUser, err := user.UpdateAUser(server.DB, uid)
 	if err != nil {
 		formattedError := formaterror.FormatError(err.Error())
 		responses.ERROR(w, http.StatusInternalServerError, formattedError)
@@ -148,7 +148,7 @@ func (server *Server) DeleteUser(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	if tokenID != 0 && tokenID != uint32(uid) {
+	if tokenID != 0 && tokenID != uid {
 		responses.ERROR(w, http.StatusUnauthorized, errors.New(http.StatusText(http.StatusUnauthorized)))
 		return
